+++ botrunner/internal/game/reward.go
@@ -4,6 +4,6 @@ type Reward struct {
 	Id       int
 	Name     string
 	Type     string
-	Amount   int64
+	Amount   float32
 	Schedule string
 }
